backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Data models and validation
pydantic==2.5.0
pandas==2.0.3

# Database dependencies
motor==3.3.2  # MongoDB async driver
pymongo==4.6.0
dnspython==2.4.2  # For MongoDB Atlas

# Authentication dependencies
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# CORS handling
starlette==0.27.0

# FREE Translation Libraries
deep-translator==1.11.4  # Multiple free translation engines
translate==3.6.1  # MyMemory API (free tier)
translators==5.7.4  # Multiple free APIs in one package
langdetect==1.0.9  # Language detection

# Alternative offline translation (smaller but free)
# argostranslate==1.9.0  # Offline translation (download models)

# HTTP client for API calls
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Caching for optimization
cachetools==5.3.2
redis==5.0.1  # Optional: For production caching

# Optional AI integration
google-generativeai==0.3.2

# Machine Learning for crop disease detection
protobuf==3.20.3
tensorflow==2.13.0
numpy==1.24.3

# Image processing
Pillow==10.1.0

# Speech recognition
SpeechRecognition==3.10.0
pyaudio==0.2.11  # For microphone input (optional)

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging and monitoring
structlog==23.2.0





